requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
mysqlclient==2.0.1
phonenumbers==8.12.10
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
